A long time ago the BuildLiveCD scripts started to rely on features from the Scripts package. That was back in the day when we had a GoboLinux host system, and we were mastering new GoboLinux releases from that host.
Now that we are in the process of releasing our first 64-bit release,
we are using another distro as host. Consequently, BuildLiveCD fails
to find the Gobo scripts.

A possible workaround is to install Rootless, but that comes with its
problems. For instance, Rootless features scripts that assume the
original filesystem layout based on /System/Links.

Therefore, this commit introduces a new script that implements
compatibility functions -- just enough to parse command line options
and to log messages. Further features can be imported from the real
Scripts package on later stages of each script that needs them.